Ensure that there is no server when noServer is true
This is a primarily a quality of life change. If you were attempting to add authentication the recommended way, you'd set `noServer` to `true`, but the fastify server is still merged and causes errors. The fix of course would be to set `server` to `null` or `undefined, however that is one extra step I did not want to do and did not want to figure out I had to do.